Added PID for light source, refactored config, new param - TotalSunTime for web page

+++ Models.h
@@ -2,7 +2,7 @@ struct Configuration {
 	float Latitude;
 	float Longitude;
 	float DesiredTemperature;
-	float DesiredLightning;
+	double DesiredLightning;
 	int CloudsSimulationPercent;
 	bool SimulateData;
 };
